Fix heuristic bug.

+++ src/heuristics/HeuristicEvaluator.java
@@ -3,5 +3,5 @@ package heuristics;
 import puzzle.PuzzleState;
 
 public interface HeuristicEvaluator {
-    public int evaluate(PuzzleState state);
+    public double evaluate(PuzzleState state);
 }
